+++ changelog.md
@@ -2,6 +2,7 @@
 
 ## 8.7.0
 
+* Added simplified `wsSelect` directive for easy `ng-select` setups
 * Added entity tree select component
 * Cross Browser date helper functions
 * userService.getUser